Crime SQL Foreign key violation resolved

+++ src/main/java/com/prison/project/service/crime/DeleteCrimeService.java
@@ -5,6 +5,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Transactional
 @Service
 @RequiredArgsConstructor
@@ -12,7 +14,14 @@ public class DeleteCrimeService {
 
     private final CrimeRepository crimeRepository;
 
-    public void deleteCrimeById(Long id){
-        crimeRepository.deleteById(id);
+    public void deleteCrimeById(Long id) {
+        List<Long> crimeList = crimeRepository.getPrisonerCrimes(id);
+        if (crimeList.isEmpty()) {
+            crimeRepository.deleteById(id);
+        }
+    }
+
+    public List<Long> getCrimePrisoner(Long id) {
+        return crimeRepository.getPrisonerCrimes(id);
     }
 }
